+++ Assets/Scripts/Game/GameRoot/Commands/HardCurrency/CommandAddHardCurrency.cs
@@ -4,7 +4,7 @@ namespace Game.GameRoot.Commands.HardCurrency
 {
     public class CommandAddHardCurrency : ICommand
     {
-        public int Value;
+        public long Value;
         
     }
 }
